requirements.txt
Flask==3.0.0
langchain==0.0.350
openai==1.3.9
python-dotenv==1.0.0